photo_label/requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain-openai
langchain-core
python-dotenv
requests
pillow

